Cattche is alpha ready
The constraints have been optimised for iPhone 6
Keyboard retracts from profile view.
The app does not crash on device
Everything is fully functional but drop down menu in maps does not
update unless it is clicked on the side which is enabled by
UItapgesture recogniser. and drop down menu when adding folders is
jumpy. this is because it is a subview of the main map view rather then
the pop up view since otherwise it does not let you do any clicks
outside when clicking in the map view.